feat(poly): use F::ROOT_OF_INFINITY in lagrange
